requirements.txt
mcp>=1.0.0
elevenlabs>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
pystray>=0.19.5
Pillow>=10.0.0
keyboard>=0.13.5
pyperclip>=1.8.2
pyautogui>=0.9.54
requests>=2.28.0
pygame-ce>=2.5.0
